+++ docs/changelog.md
@@ -3,3 +3,4 @@
 ### 2025-09-09
 
 - Add changelog for tracking user-relevant changes.
+- Add rate_limiting_levels configuration.